all 100% ready

+++ app/Http/Requests/LeadRequest.php
@@ -25,7 +25,7 @@ class LeadRequest extends FormRequest
     {
         return [
             'name' => 'string',
-            'price' => 'numeric'
+            'price' => 'integer'
         ];
     }
 }
